# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PayD Backend API",
    "version": "1.0.0",
    "description": "API Documentation for PayD Backend"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/freeze/account/freeze": {
      "post": {
        "summary": "Freeze an account's trustline",
        "tags": [
          "Freeze"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "targetAccount": {
                    "type": "string"
                  },
                  "assetCode": {
                    "type": "string"
                  },
                  "assetIssuer": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/freeze/account/unfreeze": {
      "post": {
        "summary": "Unfreeze an account's trustline",
        "tags": [
          "Freeze"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "targetAccount": {
                    "type": "string"
                  },
                  "assetCode": {
                    "type": "string"
                  },
                  "assetIssuer": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/freeze/global/freeze": {
      "post": {
        "summary": "Pause transfers globally for an asset",
        "tags": [
          "Freeze"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assetCode": {
                    "type": "string"
                  },
                  "assetIssuer": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/freeze/global/unfreeze": {
      "post": {
        "summary": "Restore transfers globally for an asset",
        "tags": [
          "Freeze"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "assetCode": {
                    "type": "string"
                  },
                  "assetIssuer": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/freeze/status/{targetAccount}": {
      "get": {
        "summary": "Query the active freeze status of an account",
        "tags": [
          "Freeze"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "targetAccount",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "assetCode",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "assetIssuer",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/freeze/logs": {
      "get": {
        "summary": "History of all freeze and unfreeze actions",
        "tags": [
          "Freeze"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "targetAccount",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "assetCode",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}